fix detekt issue on GridComposable
