+++ components/remote-controls/grid/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/grid/composable/preview/LoadedContentPreview.kt
@@ -3,7 +3,7 @@ package com.flipperdevices.remotecontrols.impl.grid.composable.preview
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import com.flipperdevices.core.ui.theme.FlipperThemeInternal
-import com.flipperdevices.remotecontrols.impl.grid.composable.LoadedContent
+import com.flipperdevices.remotecontrols.impl.grid.composable.components.GridComposableLoadedContent
 
 @Preview(
     showSystemUi = true,
@@ -12,7 +12,7 @@ import com.flipperdevices.remotecontrols.impl.grid.composable.LoadedContent
 @Composable
 private fun LoadedContentPreview() {
     FlipperThemeInternal {
-        LoadedContent(
+        GridComposableLoadedContent(
             pagesLayout = KitchenLayoutFactory.create(),
             onButtonClick = { _, _ -> },
             onReload = {},
